Fix integer sizing issues in the DiskIO Meter
On Linux kernels the size of the values exported for block
device bytes has used a 64 bit integer for quite some time
(2.6+ IIRC).  Make the procfs value extraction use correct
types and change internal types used to rate convert these
counters (within the DiskIO Meter) 64 bit integers, where
appropriate.

+++ DiskIOMeter.h
@@ -10,9 +10,9 @@ in the source distribution for its full text.
 #include "Meter.h"
 
 typedef struct DiskIOData_ {
-   unsigned long int totalBytesRead;
-   unsigned long int totalBytesWritten;
-   unsigned long int totalMsTimeSpend;
+   unsigned long long int totalBytesRead;
+   unsigned long long int totalBytesWritten;
+   unsigned long long int totalMsTimeSpend;
 } DiskIOData;
 
 extern const MeterClass DiskIOMeter_class;
